kafka_producer/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
dataclasses-json==0.6.1
python-dotenv==1.0.0
